# perf: improve scale performance by using min/max directly instead of range/diff